lifted state up to App.js and forwaded it to UserInput usinh props
