+++ src/components/Results.jsx
@@ -1,0 +1,3 @@
+export default function Results() {
+    return <p>results...</p>;
+}
