+++ src/components/UserInput.jsx
@@ -1,22 +1,4 @@
-import { useState } from "react";
-
-export default function UserInput() {
-    const [userInput, setUserInput] = useState({
-        initialInvestment: 10000,
-        annualInvestment: 1200,
-        expectedReturn: 6,
-        duration: 10,
-    });
-
-    function handleChange(inputIdentifier, newValue) {
-        setUserInput((prevUserInput) => {
-            return {
-                ...prevUserInput,
-                [inputIdentifier]: newValue,
-            };
-        });
-    }
-
+export default function UserInput({ onChange, userInput }) {
     return (
         <section id="user-input">
             <div className="input-group">
@@ -26,7 +8,7 @@ export default function UserInput() {
                         type="number"
                         value={userInput.initialInvestment}
                         onChange={(event) =>
-                            handleChange("initialInvestment", event.target.value)
+                            onChange("initialInvestment", event.target.value)
                         }
                     />
                 </p>
@@ -37,7 +19,7 @@ export default function UserInput() {
                         required
                         value={userInput.annualInvestment}
                         onChange={(event) =>
-                            handleChange("annualInvestment", event.target.value)
+                            onChange("annualInvestment", event.target.value)
                         }
                     />
                 </p>
@@ -49,9 +31,7 @@ export default function UserInput() {
                         type="number"
                         required
                         value={userInput.expectedReturn}
-                        onChange={(event) =>
-                            handleChange("expectedReturn", event.target.value)
-                        }
+                        onChange={(event) => onChange("expectedReturn", event.target.value)}
                     />
                 </p>
                 <p>
@@ -60,7 +40,7 @@ export default function UserInput() {
                         type="number"
                         required
                         value={userInput.duration}
-                        onChange={(event) => handleChange("duration", event.target.value)}
+                        onChange={(event) => onChange("duration", event.target.value)}
                     />
                 </p>
             </div>
